toyproject/requirements.txt
asgiref==3.7.2
attrs==23.2.0
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.85
cement==2.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.3
Django==5.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
idna==3.6
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
pathspec==0.10.1
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.16.2
semantic-version==2.8.5
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.1.9
